+++ be/src/main/java/kr/codesquad/secondhand/api/chat/service/ChatRoomFacadeService.java
@@ -7,9 +7,9 @@ import java.util.Optional;
 import kr.codesquad.secondhand.api.chat.domain.ChatMessage;
 import kr.codesquad.secondhand.api.chat.domain.ChatRoom;
 import kr.codesquad.secondhand.api.chat.dto.ChatRoomCreateDto;
-import kr.codesquad.secondhand.api.chat.dto.reponse.ChatRoomExistenceCheckResponse;
-import kr.codesquad.secondhand.api.chat.dto.reponse.ChatRoomMessagesReadResponse;
-import kr.codesquad.secondhand.api.chat.dto.reponse.ChatRoomReadResponse;
+import kr.codesquad.secondhand.api.chat.dto.response.ChatRoomExistenceCheckResponse;
+import kr.codesquad.secondhand.api.chat.dto.response.ChatRoomMessagesReadResponse;
+import kr.codesquad.secondhand.api.chat.dto.response.ChatRoomReadResponse;
 import kr.codesquad.secondhand.api.member.domain.Member;
 import kr.codesquad.secondhand.api.member.service.MemberService;
 import kr.codesquad.secondhand.api.product.domain.Product;
